requirements.txt
python-dotenv
logging
asyncio
aiohttp
aiogram
huggingface_hub